code_analyzer_backend/requirements.txt
pyyaml
python-dotenv
requests
fastapi[standard]
